+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Add controllers, services and repositories unit tests ([#6](https://github.com/nahuelsoma/movies-api/pull/6))
 - Add schema validation ([#9](https://github.com/nahuelsoma/movies-api/pull/9))
 - Enhace error logging ([#8](https://github.com/nahuelsoma/movies-api/pull/8))
 - Enhace error handling ([#7](https://github.com/nahuelsoma/movies-api/pull/7))